+++ example/android/app/src/main/kotlin/io/github/mertguner/sound_generator_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.github.mertguner.sound_generator_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()